% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_rep.R
\name{auto_rep}
\alias{auto_rep}
\title{auto_rep}
\usage{
auto_rep(
  counts,
  genes_id,
  metadata,
  muts,
  response,
  design,
  colors = c("black", "orange"),
  ref_level,
  shrink = "apeglm",
  biomart,
  fold_change = 2,
  p.adj = 0.05,
  gmt,
  gsea_pvalue = 0.2,
  gsva_gmt = "hallmark",
  method = "gsva",
  row.names = TRUE,
  col.names = TRUE,
  top_genes = 10,
  specific_genes = NULL,
  compare = "wilcox.test",
  p_label = "p.format",
  gbuild,
  mut_sigs,
  tri.counts.method = "default",
  indications,
  cibersort = NULL,
  tumor = TRUE,
  rmgenes = NULL,
  scale_mrna = TRUE,
  expected_cell_types = NULL
)
}
\arguments{
\item{expected_cell_types}{}
}
\value{

}
\description{
auto_rep
}
\examples{
prova(counts = input_ge_module,
      genes_id = 'entrezgene_id',
      metadata = met,
      response = 'Response',
      design = 'Response',
      colors = c('black', 'orange', 'purple'),
      ref_level = c('Response', 'Non_Responders'),
      shrink = 'none',
      biomart = ensembl_biomart_GRCh38_p13,
      fold_change = 2,
      p.adj = 0.05,
      gmt = 'c7.all.v7.2.symbols.gmt',
      gsea_pvalue = 0.2,
      gsva_gmt = 'hallmark',
      method = 'gsva',
      row.names = TRUE,
      col.names = TRUE,
      muts = input_gv_module,
      top_genes = 10,
      specific_genes = NULL,
      compare = 'wilcox.test',
      p_label = 'p.format',
      gbuild = 'BSgenome.Hsapiens.UCSC.hg19',
      mut_sigs = 'COSMIC_v2_SBS_GRCh37',
      tri.counts.method = 'default',
      indications = rep('skcm', ncol(input_ge_module[-1])),
      cibersort = NULL,
      tumor = TRUE,
      rmgenes = NULL,
      scale_mrna = TRUE,
      expected_cell_types = NULL)

}
