% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ge_annot.R
\name{ge_annot}
\alias{ge_annot}
\title{ge_annot}
\usage{
ge_annot(results_dds, genes_id, biomart)
}
\arguments{
\item{results_dds}{The output of the ge_diff_exp function. List containing
data frames of differential gene expression results between the different groups.}

\item{genes_id}{Name of the column that contains gene identifiers. Should be
one of the following: 'ensembl_gene_id', 'entrezgene_id' or 'hgnc_symbol'.}

\item{biomart}{Data frame containing a biomaRt query with the following
attributes: ensembl_gene_id, hgnc_symbol, entrezgene_id, transcript_length,
refseq_mrna.}
}
\value{
Returns a list of data frames containing differential gene expression
results, one for each level comparison where genes have been annotated using
HGNC symbols.
}
\description{
Associates Entrez_gene_id (NCBI-ID) or ENSEMBL_gene_id to
HGNC symbols (Hugo Gene Nomenclature Committee nomenclature).
}
\examples{
results_dds <- ge_diff_exp(counts = input_ge_module,
                           genes_id = 'entrezgene_id',
                           metadata = metadata_ge_module,
                           design = 'Response',
                           ref_level = c('Response', 'Non_Responders'),
                           shrink = 'apeglm')
annot.res <- ge_annot(results_dds = results_dds,
                      genes_id = 'entrezgene_id',
                      biomart = ensembl_biomart_GRCh38_p13)
}
