% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ge_diff_exp.R
\name{ge_diff_exp}
\alias{ge_diff_exp}
\title{ge_diff_exp}
\usage{
ge_diff_exp(counts, genes_id, metadata, design, ref_level)
}
\arguments{
\item{counts}{Data frame that contains gene expression data as raw counts.}

\item{genes_id}{Name of the column that contains gene identifiers. Should be
one of the following:'entrez_gene_id', 'ensembl_gene_id' or 'hgnc_symbol'.}

\item{metadata}{Data frame that contains supporting variables to the data.}

\item{design}{Variables in the design formula in the form of: 'Var1 + Var2 + ... Var_n'.}

\item{ref_level}{Character vector where the first element is the column name
where the reference level is located and a second element indicating the name
of level to be used as a reference when calculating differential gene
expression.}
}
\value{
Returns a list of differential gene expression results, one for each
level comparison, in form of DESeqResults objects.
}
\description{

}
\examples{
results_dds <- ge_diff_exp(counts = input_ge_module,
genes_id = 'entrezgene_id', metadata = metadata_ge_module,
design = 'Response', ref_level = c('Response', 'Non_Responders'))

}
