% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ge_gsea.R
\name{ge_gsea}
\alias{ge_gsea}
\title{ge_gsea}
\usage{
ge_gsea(annot_res, gmt, gsea_pvalue = 0.2)
}
\arguments{
\item{annot_res}{The output of ge_annot. List containing data frames of
differential gene expression results between the different groups.}

\item{gmt}{A data frame containg the gene sets to analyse usgin GSEA. This
object should be obtained using the read.gmt function from the clusterProfiler
package.}

\item{gsea_pvalue}{Numeric value to define the adjusted pvalue cutoff during
GSEA. Set to 0.2 by default.}
}
\value{

}
\description{

}
\examples{

}
