% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gv_extr_mut_sig.R
\name{gv_extr_mut_sig}
\alias{gv_extr_mut_sig}
\title{gv_extr_mut_sig}
\usage{
gv_extr_mut_sig(results, ids_samples)
}
\arguments{
\item{results}{Data frame containing the results of whichSignatures function
from the deconstructSigs package.}

\item{ids_samples}{Name of the samples of interest.}
}
\value{
Returns a data frame with the contribution of each mutational signature
studied in each sample in a long format.
}
\description{
Extracts mutational signatures contribution prediction from the
deconstrucSigs package in a legible format.
}
\examples{
mut.sbs <- input_gv_module \%>\% dplyr::filter(Variant_Type == 'SNP')
sigs.sbs.input <- deconstructSigs::mut.to.sigs.input(
                                          mut.ref = mut.sbs,
                                          sample.id = 'Tumor_Sample_Barcode',
                                          chr = 'Chromosome',
                                          pos = 'Start_Position',
                                          ref = 'Reference_Allele',
                                          alt = 'Tumor_Seq_Allele2',
                                          bsg = BSgenome.Hsapiens.UCSC.hg19,
                                          sig.type = 'SBS')
ids_samples <- unique(input_gv_module$Tumor_Sample_Barcode)
results_sbs <- sapply(ids_samples,
                      function(x) {
                         deconstructSigs::whichSignatures(
                         tumor.ref = sigs.sbs.input,
                         signatures.ref = signatures.cosmic,
                         sample.id = x,
                         contexts.needed = TRUE,
                         tri.counts.method = 'default')
                     })

results_sbs.extr <- gv_extr_mut_sig(results_sbs)

}
