% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gv_mut_load.R
\name{gv_mut_load}
\alias{gv_mut_load}
\title{gv_mut_load}
\usage{
gv_mut_load(
  muts,
  metadata,
  response,
  compare = NULL,
  p_label = "p.format",
  colors = c("black", "orange")
)
}
\arguments{
\item{colors}{}
}
\value{

}
\description{

}
\examples{
gv_mut_load(muts = input_gv_module,
            metadata = metadata_ge_module,
            response = Response,
            compare = 'wilcox.test',
            p_label = 'p.format',
            colors = c('black', 'orange'))

}
