% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gv_mut_signatures.R
\name{gv_mut_signatures}
\alias{gv_mut_signatures}
\title{gv_mut_signatures}
\usage{
gv_mut_signatures(
  muts,
  metadata,
  response,
  gbuild = "BSgenome.Hsapiens.UCSC.hg19",
  mut_sigs = "COSMIC_v2_SBS_GRCh37",
  tri.counts.method = "default",
  colors = c("black", "orange"),
  col.names = TRUE
)
}
\arguments{
\item{muts}{Data frame containing genomic variations. Necessary columns must
have the following names:
- Hugo_Symbol: Gene symbol from HGNC.
- Chromosome: Affected chromosome.
- Start_Position: Mutation start coordinate.
- End_Position: Mutation end coordinate.
- Reference_Allele: The plus strand reference allele at this position.
Includes the deleted sequence for a deletion or "-" for an insertion.
- Tumor_Seq_Allele2: Tumor sequencing discovery allele.
- Variant_Classification: Translational effect of variant allele. Can be one
of the following: Frame_Shift_Del, Frame_Shift_Ins, In_Frame_Del,
In_Frame_Ins, Missense_Mutation, Nonsense_Mutation, Silent, Splice_Site,
Translation_Start_Site, Nonstop_Mutation, RNA, Targeted_Region.
- Variant_Type: Type of mutation. Can be: 'SNP' (Single nucleotide polymorphism),
'DNP' (Double nucleotide polymorphism), 'INS' (Insertion), 'DEL' (Deletion).
- Tumor_Sample_Barcode: Sample name.}

\item{metadata}{Data frame that contains supporting variables to the data.}

\item{response}{Unquoted name of the variable indicating the groups to analyse.}

\item{gbuild}{Version of the genome to work with. It can be one of the following:
- ‘BSgenome.Hsapiens.UCSC.hg19’
- ‘BSgenome.Hsapiens.UCSC.hg38’
- ‘BSgenome.Mmusculus.UCSC.mm10’
- ‘BSgenome.Mmusculus.UCSC.mm39’}

\item{mut_sigs}{Mutational signature matrices containing the frequencies of
all nucleotide changes per signature need to be indicated. GEGVIC contains the
matrices from COSMIC for single and double base substitutions. To choose one,
the user has to indicate ’COSMIC_v{XX}_{YY}BS_GRCh{ZZ}’ in the mut_sigs argument.
The XX is the version, that can be v2 or v3.2. YY indicates if mutations are
single (S) or double (D) base substitutions, while the ZZ is for the genome
assembly, either GRCh37 or GRCh38 for human data and mm9 or mm10 for mouse data.}

\item{tri.counts.method}{Normalization method. Needs to be set to either:
- 'default' – no further normalization.
- 'exome' – normalized by number of times each trinucleotide context is
observed in the exome.
- 'genome' – normalized by number of times each trinucleotide context is
observed in the genome.
- 'exome2genome' – multiplied by a ratio of that trinucleotide's occurence in
the genome to the trinucleotide's occurence in the exome.
- 'genome2exome' – multiplied by a ratio of that trinucleotide's occurence in
the exome to the trinucleotide's occurence in the genome.
- data frame containing user defined scaling factor – count data for each
trinucleotide context is multiplied by the corresponding value given in the
data frame.}

\item{colors}{Character vector indicating the colors of the different groups
to compare. Default values are two: black and orange.}

\item{col.names}{Logical value to determine if tumour are shown in plots.}
}
\value{
Returns ggplot objects and the results table in a form of a data frame.
}
\description{
Predicts the contribution of known mutational processes to the samples.
}
\examples{
mut.sigs <- gv_mut_signatures(muts = sample_mutations,
                              metadata = sample_metadata,
                              response = MSI_status,
                              gbuild = 'BSgenome.Hsapiens.UCSC.hg38',
                              mut_sigs = 'COSMIC_v2_SBS_GRCh38',
                              tri.counts.method = 'default',
                              colors = c('orange', 'black'),
                              col.names = TRUE)

}
