% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gv_mut_summary.R
\name{gv_mut_summary}
\alias{gv_mut_summary}
\title{gv_mut_summary}
\usage{
gv_mut_summary(
  df,
  metadata,
  response,
  top_genes = 10,
  specific_genes = NULL,
  colors = c("black", "orange")
)
}
\arguments{
\item{colors}{}
}
\value{

}
\description{
gv_mut_summary
}
