% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic_plot_comp_celltypes.R
\name{ic_plot_comp_celltypes}
\alias{ic_plot_comp_celltypes}
\title{ic_plot_comp_celltypes}
\usage{
ic_plot_comp_celltypes(df, metadata, response)
}
\arguments{
\item{response}{}
}
\value{
Returns a ggplot object.
}
\description{

}
\examples{



}
