% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic_plot_comp_samples.R
\name{ic_plot_comp_samples}
\alias{ic_plot_comp_samples}
\title{ic_plot_comp_samples}
\usage{
ic_plot_comp_samples(
  df,
  metadata,
  response,
  compare = NULL,
  p_label = "p.format",
  colors = c("black", "orange")
)
}
\arguments{
\item{colors}{}
}
\value{
Returns a ggplot object.
}
\description{

}
\examples{





}
