% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic_raw_to_tpm.R
\name{ic_raw_to_tpm}
\alias{ic_raw_to_tpm}
\title{ic_raw_to_tpm}
\usage{
ic_raw_to_tpm(counts, genes_id, biomart)
}
\arguments{
\item{counts}{Data frame that contains gene expression data as raw counts.}

\item{genes_id}{Name of the column that contains gene identifiers. Should be
one of the following:'entrez_gene_id', 'ensemblgene_id' or 'hgnc_symbol'.}

\item{biomart}{Data frame containing a biomaRt query with the following
attributes: ensembl_gene_id, hgnc_symbol, entrezgene_id, transcript_length,
refseq_mrna.}
}
\value{
Returns a matrix containing expression counts as TPM with HGNC gene
symbols as rownames and samples identifiers as colnames.
}
\description{
Transforms RNA-seq raw counts to TPM (Transcript Per kilobase
Million).
}
\examples{
tpm <- ic_raw_to_tpm(counts = input_ge_module,
                     genes_id = 'entrezgene_id',
                     biomart = ensembl_biomart_GRCh38_p13)
}
