% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic_score.R
\name{ic_score}
\alias{ic_score}
\title{ic_score}
\usage{
ic_score(
  tpm,
  metadata,
  response,
  compare = NULL,
  p_label = "p.format",
  colors = c("black", "orange")
)
}
\arguments{
\item{colors}{}
}
\value{

}
\description{

}
\examples{



}
