% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_ic.R
\name{module_ic}
\alias{module_ic}
\title{module_ic}
\usage{
module_ic(
  counts,
  genes_id,
  biomart,
  indications = NULL,
  cibersort = NULL,
  tumor = TRUE,
  rmgenes = NULL,
  scale_mrna = TRUE,
  expected_cell_types = NULL,
  metadata,
  response,
  compare = NULL,
  p_label = "p.format",
  colors = c("black", "orange")
)
}
\arguments{
\item{colors}{}
}
\value{

}
\description{
module_ic
}
