% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ge_data_preprocess.R
\name{preprocess_ge_meta}
\alias{preprocess_ge_meta}
\title{preprocess_ge_meta}
\usage{
preprocess_ge_meta(metadata)
}
\arguments{
\item{metadata}{Data frame that contains supporting variables to the data}
}
\value{
Generates a data frame.
}
\description{

}
\examples{

}
