% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ge_annot.R
\name{ge_annot}
\alias{ge_annot}
\title{Title}
\usage{
ge_annot(results_dds, genes_id)
}
\arguments{
\item{genes_id}{Name of the column that contains gene identifiers. Should be
one of the following:'entrez_gene_id', 'ensembl_gene_id' or 'hgnc_symbol'.}
}
\value{

}
\description{
Title
}
\examples{
ge_annot(results_dds = results_dds, genes_id = 'entrezgene_id')
}
