% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ge_annot.R
\name{ge_annot}
\alias{ge_annot}
\title{ge_annot}
\usage{
ge_annot(results_dds, genes_id, biomart)
}
\arguments{
\item{results_dds}{The output of the ge_diff_exp function. List containing
data frames of differential gene expression results between the different groups.}

\item{genes_id}{Name of the column that contains gene identifiers. Should be
one of the following: 'ensembl_gene_id', 'entrezgene_id' or 'hgnc_symbol'.}

\item{biomart}{Data frame containing a biomaRt query with the following
attributes: ensembl_gene_id, hgnc_symbol, entrezgene_id, transcript_length,
refseq_mrna. In the case of mus musculus data, external_gene_name must be
obtained and then change the column name for hgnc_symbol. Uploaded biomaRt
queries in GEGVIC: 'ensembl_biomartGRCh37', ensembl_biomartGRCh38_p13' and
'ensembl_biomartGRCm38_p6', 'ensembl_biomartGRCm39'.}
}
\value{
Returns a list of data frames containing differential gene expression
results, one for each level comparison where genes have been annotated using
HGNC symbols.
}
\description{
Associates Entrez_gene_id (NCBI-ID) or ENSEMBL_gene_id to
HGNC symbols (Hugo Gene Nomenclature Committee nomenclature).
}
\examples{
results.dds <- ge_diff_exp(counts = sample_counts,
                           genes_id = 'ensembl_gene_id',
                           metadata = sample_metadata,
                           design = 'MSI_status',
                           ref_level = c('MSI_status', 'MSS'),
                           shrink = 'apeglm')
annot.res <- ge_annot(results_dds = results_dds,
                      genes_id = 'ensembl_gene_id',
                      biomart = ensembl_biomart_GRCh38_p13)
}
