% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ge_gsea.R
\name{ge_gsea}
\alias{ge_gsea}
\title{ge_gsea}
\usage{
ge_gsea(annot_res, gmt, gsea_pvalue = 0.2)
}
\arguments{
\item{annot_res}{The output of ge_annot. List containing data frames of
differential gene expression results between the different groups.}

\item{gmt}{Path to the gmt file that contain the gene sets of interest.}

\item{gsea_pvalue}{Numeric value to define the adjusted pvalue cutoff during
GSEA. Set to 0.2 by default.}
}
\value{
Returns ggplot objects and a list of gseaResult objects.
}
\description{
Performs Gene Set Enrichment Analysis using gene sets specified by the user.
}
\examples{
results.dds <- ge_diff_exp(counts = sample_counts,
                           genes_id = 'ensembl_gene_id',
                           metadata = sample_metadata,
                           design = 'MSI_status',
                           ref_level = c('MSI_status', 'MSS'),
                           shrink = 'apeglm')
annot.res <- ge_annot(results_dds = results_dds,
                      genes_id = 'ensembl_gene_id',
                      biomart = ensembl_biomart_GRCh38_p13)
gsea.res <- ge_gsea(annot_res = annot.res,
                    gmt = 'inst/extdata/c2.cp.reactome.v7.5.1.symbols.gmt',
                    gsea_pvalue = 0.2)

}
