% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ge_gsea.R
\name{ge_gsea}
\alias{ge_gsea}
\title{ge_gsea}
\usage{
ge_gsea(annot_res, gmt, gsea_pvalue = 0.2)
}
\arguments{
\item{annot_res}{The output of ge_annot. List containing data frames of
differential gene expression results between the different groups.}

\item{gmt}{Path to the gmt file that contain the gene sets of interest.}

\item{gsea_pvalue}{Numeric value to define the adjusted pvalue cutoff during
GSEA. Set to 0.2 by default.}
}
\value{
Returns ggplot objects and a list of gseaResult objects.
}
\description{

}
\examples{
results_dds <- ge_diff_exp(counts = input_ge_module,
                           genes_id = 'entrezgene_id',
                           metadata = metadata_ge_module,
                           design = 'Response',
                           ref_level = c('Response', 'Non_Responders'),
                           shrink = 'none')
annot.res <- ge_annot(results_dds = results_dds,
                      genes_id = 'entrezgene_id',
                      biomart = ensembl_biomart_GRCh38_p13)
gsea.res <- ge_gsea(annot_res = annot.res,
                    gmt = 'inst/extdata/c7.all.v7.2.symbols.gmt',
                    gsea_pvalue = 0.2)

}
