% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ge_pca.R
\name{ge_pca}
\alias{ge_pca}
\title{Title}
\usage{
ge_pca(data, gs, metadata, group, colors = c("black", "orange"))
}
\arguments{
\item{data}{Data frame that contains the gene expression data}

\item{metadata}{Data frame that contains supporting variables to the data
such as sample groups, treatments, etc.}

\item{group}{Name of the column in the metadata where the grouping factor is
located.}

\item{colors}{Character vector indicating the colors of the different groups
to compare. Default values are black and orange}
}
\value{

}
\description{
Title
}
\examples{

}
