% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ge_volcano.R
\name{ge_volcano}
\alias{ge_volcano}
\title{ge_volcano}
\usage{
ge_volcano(annot_res, fold_change = 2, p.adj = 0.05)
}
\arguments{
\item{annot_res}{The output of ge_annot. List containing data frames of
differential gene expression results between the different groups.}

\item{fold_change}{An integer to define the fold change value to consider
that a gene is differentially expressed.}

\item{p.adj}{Numeric value to define the maximum adjusted p-value to consider
that a gene is differentially expressed.}
}
\value{
Returns ggplot graphs.
}
\description{
Plots volcano plots from the outputs of the ge_annot() function.
}
\examples{
results_dds <- ge_diff_exp(counts = input_ge_module,
                           genes_id = 'entrezgene_id',
                           metadata = metadata_ge_module,
                           design = 'Response',
                           ref_level = c('Response', 'Non_Responders'),
                           shrink = 'apeglm')
annot.res <- ge_annot(results_dds = results_dds, genes_id = 'entrezgene_id')
ge_volcano(annot_res = annot.results, fold_change = 2, p.adj = 0.05)
}
