% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gv_mut_load.R
\name{gv_mut_load}
\alias{gv_mut_load}
\title{gv_mut_load}
\usage{
gv_mut_load(
  muts,
  metadata,
  response,
  compare = NULL,
  p_label = "p.format",
  colors = c("black", "orange")
)
}
\arguments{
\item{muts}{Data frame containing genetic variations. Necessary columns must
have the following names:
- Hugo_Symbol: Gene symbol from HGNC.
- Chromosome: Affected chromosome.
- Start_Position: Mutation start coordinate.
- End_Position: Mutation end coordinate.
- Reference_Allele: The plus strand reference allele at this position.
Includes the deleted sequence for a deletion or "-" for an insertion.
- Tumor_Seq_Allele2: Tumor sequencing discovery allele.
- Variant_Classification: Translational effect of variant allele. Can be one
of the following: Frame_Shift_Del, Frame_Shift_Ins, In_Frame_Del,
In_Frame_Ins, Missense_Mutation, Nonsense_Mutation, Silent, Splice_Site,
Translation_Start_Site, Nonstop_Mutation, RNA, Targeted_Region.
- Variant_type: Type of mutation. Can be: 'SNP' (Single nucleotide polymorphism),
'DNP' (Double nucleotide polymorphism), 'INS' (Insertion), 'DEL' (Deletion).
- Tumor_Sample_Barcode: Sample name.}

\item{metadata}{Data frame that contains supporting variables to the data.}

\item{response}{Unquoted name of the variable indicating the groups to analyse.}

\item{compare}{A character string indicating which method to be used for
comparing means. Options are 't.test' and 'wilcox.test' for two groups or
'anova' and 'kruskal.test' for more groups. Default value is NULL.}

\item{p_label}{Character string specifying label type. Allowed values include
'p.signif' (shows the significance levels), 'p.format' (shows the formatted
p-value).}

\item{colors}{Character vector indicating the colors of the different groups
to compare. Default values are two: black and orange.}
}
\value{
Returns a ggplot object.
}
\description{
Summarises the total number of mutations per sample and compares
different groups of interest.
}
\examples{
gv_mut_load(muts = input_gv_module,
            metadata = metadata_ge_module,
            response = Response,
            compare = 'wilcox.test',
            p_label = 'p.format',
            colors = c('black', 'orange'))

}
