% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gv_mut_signatures.R
\name{gv_mut_signatures}
\alias{gv_mut_signatures}
\title{Title}
\usage{
gv_mut_signatures(
  muts,
  metadata,
  response,
  gbuild = "BSgenome.Hsapiens.UCSC.hg19",
  mut_sigs = COSMIC_v2_SBS_GRCh37,
  tri.counts.method = "default",
  colors = c("black", "orange")
)
}
\arguments{
\item{colors}{}
}
\value{

}
\description{
Title
}
\examples{

gv_mut_signatures(muts = input_gv_module,
                  metadata = metadata_ge_module,
                  response = Response,
                  gbuild = 'BSgenome.Hsapiens.UCSC.hg19',
                  mut_sigs = COSMIC_v2_SBS_GRCh37,
                  tri.counts.method = 'default',
                  colors = c('black', 'orange'))

}
