% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gv_mut_summary.R
\name{gv_mut_summary}
\alias{gv_mut_summary}
\title{gv_mut_summary}
\usage{
gv_mut_summary(
  muts,
  metadata,
  response,
  top_genes = 10,
  specific_genes = NULL,
  colors = c("black", "orange")
)
}
\arguments{
\item{colors}{}
}
\value{

}
\description{
gv_mut_summary
}
\examples{
gv_mut_summary(muts = input_gv_module,
               metadata = metadata_ge_module,
               response = Response,
               top_genes = 10,
               specific_genes = NULL,
               colors = c('black', 'orange'))


}
