% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gv_mut_summary.R
\name{gv_mut_summary}
\alias{gv_mut_summary}
\title{gv_mut_summary}
\usage{
gv_mut_summary(
  muts,
  metadata,
  response,
  top_genes = 10,
  specific_genes = NULL,
  col.names = TRUE,
  colors = c("black", "orange")
)
}
\arguments{
\item{muts}{Data frame containing genetic variations. Necessary columns must
have the following names:
- Hugo_Symbol: Gene symbol from HGNC.
- Chromosome: Affected chromosome.
- Start_Position: Mutation start coordinate.
- End_Position: Mutation end coordinate.
- Reference_Allele: The plus strand reference allele at this position.
Includes the deleted sequence for a deletion or "-" for an insertion.
- Tumor_Seq_Allele2: Tumor sequencing discovery allele.
- Variant_Classification: Translational effect of variant allele. Can be one
of the following: Frame_Shift_Del, Frame_Shift_Ins, In_Frame_Del,
In_Frame_Ins, Missense_Mutation, Nonsense_Mutation, Silent, Splice_Site,
Translation_Start_Site, Nonstop_Mutation, RNA, Targeted_Region.
- Variant_Type: Type of mutation. Can be: 'SNP' (Single nucleotide polymorphism),
'DNP' (Double nucleotide polymorphism), 'INS' (Insertion), 'DEL' (Deletion).
- Tumor_Sample_Barcode: Sample name.}

\item{metadata}{Data frame that contains supporting variables to the data.}

\item{response}{Unquoted name of the variable indicating the groups to analyse.}

\item{top_genes}{Number of genes to be analysed in the mutational summary.}

\item{specific_genes}{Genes that will be plotted in the oncoplot.}

\item{col.names}{Logical value to decide if tumour names are added to the plot.}

\item{colors}{Character vector indicating the colors of the different groups
to compare. Default values are two: black and orange.}
}
\value{
Prints two plots: A summary of sample mutations and an oncoplot.
}
\description{
Given a table that contains genetic variants from samples of
interest, it summarises different aspects of such mutations and draws an
oncoplot.
}
\examples{
gv_mut_summary(muts = input_gv_module,
               metadata = metadata_ge_module,
               response = Response,
               top_genes = 10,
               specific_genes = NULL,
               col.names = TRUE,
               colors = c('black', 'orange'))


}
