% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic_plot_comp_celltypes.R
\name{ic_plot_comp_celltypes}
\alias{ic_plot_comp_celltypes}
\title{ic_plot_comp_celltypes}
\usage{
ic_plot_comp_celltypes(df, metadata, response)
}
\arguments{
\item{df}{Output ot ic_deconv function. A data frame containing the predictions
of the six methods included in the immunedeconv package. First column must
indicate and be named as cell_type, method and the rest being the numeric
prediction for each sample.}

\item{metadata}{Data frame that contains supporting variables to the data.}

\item{response}{Unquoted name of the variable indicating the groups to analyse.}
}
\value{
Returns a ggplot object.
}
\description{
Plots immune cell fractions to be compared within each sample.
Results from three different methods (CIBERSORT, EPIC and QUANTISEQ) are
presented by each group of samples.
}
\examples{
tpm <- ic_raw_to_tpm(counts = sample_counts,
                     genes_id = 'ensembl_gene_id',
                     biomart = ensembl_biomart_GRCh38_p13)
ic.pred <- ic_deconv(gene_expression = tpm,
                     indications = rep('coad', ncol(tpm)),
                     cibersort = NULL,
                     tumor = TRUE,
                     rmgenes = NULL,
                     scale_mrna = TRUE,
                     expected_cell_types = NULL)
ic_plot_comp_celltypes(df = ic.pred,
                     metadata = sample_metadata,
                     response = MSI_status)

}
