% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic_plot_comp_samples.R
\name{ic_plot_comp_samples}
\alias{ic_plot_comp_samples}
\title{ic_plot_comp_samples}
\usage{
ic_plot_comp_samples(
  df,
  metadata,
  response,
  compare = NULL,
  p_label = "p.format",
  colors = c("black", "orange"),
  points = TRUE
)
}
\arguments{
\item{df}{Output ot ic_deconv function. A data frame containing the predictions
of the six methods included in the immunedeconv package. First column must
indicate and be named as cell_type, method and the rest being the numeric
prediction for each sample.}

\item{metadata}{Data frame that contains supporting variables to the data.}

\item{response}{Unquoted name of the variable indicating the groups to analyse.}

\item{compare}{A character string indicating which method to be used for
comparing means. Options are 't.test' and 'wilcox.test' for two groups or
'anova' and 'kruskal.test' for more groups. Default value is NULL.}

\item{p_label}{Character string specifying label type. Allowed values include
'p.signif' (shows the significance levels), 'p.format' (shows the formatted
p-value).}

\item{colors}{Character vector indicating the colors of the different groups
to compare. Default values are two: black and orange.}

\item{points}{Logical value to decide if points are added to the plot.}
}
\value{
Returns list containing two ggplot objects.
}
\description{
Plots predicted immune cell populations to be compared between
sample groups. Among different immune cells, the following are represented:
B cells, Macrophages, myeloid Dendritic cells, Neutrophils, Natural Killer,
T cell CD4+ and T cell CD8+.
}
\examples{
tpm <- ic_raw_to_tpm(counts = sample_counts,
                     genes_id = 'ensembl_gene_id',
                     biomart = ensembl_biomart_GRCh38_p13)
ic.pred <- ic_deconv(gene_expression = tpm,
                     indications = rep('coad', ncol(tpm)),
                     cibersort = NULL,
                     tumor = TRUE,
                     rmgenes = NULL,
                     scale_mrna = TRUE,
                     expected_cell_types = NULL)
ic_plot_comp_samples(df = ic.pred,
                     metadata = sample_metadata,
                     response = MSI_status,
                     compare = 'wilcox.test',
                     p_label = 'p.format',
                     colors = c('orange', 'black'),
                     points = TRUE)
}
