% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic_raw_to_tpm.R
\name{ic_raw_to_tpm}
\alias{ic_raw_to_tpm}
\title{ic_raw_to_tpm}
\usage{
ic_raw_to_tpm(counts, genes_id, biomart)
}
\arguments{
\item{counts}{Data frame that contains gene expression data as raw counts.}

\item{genes_id}{Name of the column that contains gene identifiers. Should be
one of the following:'entrezgene_id', 'ensembl_gene_id' or 'hgnc_symbol'.}

\item{biomart}{Data frame containing a biomaRt query with the following
attributes: ensembl_gene_id, hgnc_symbol, entrezgene_id, transcript_length,
refseq_mrna. In the case of mus musculus data, external_gene_name must be
obtained and then change the column name for hgnc_symbol. Uploaded biomaRt
queries in GEGVIC: 'ensembl_biomartGRCh37', ensembl_biomartGRCh38_p13' and
'ensembl_biomartGRCm38_p6', 'ensembl_biomartGRCm39'.}
}
\value{
Returns a matrix containing expression counts as TPM with HGNC gene
symbols as rownames and samples identifiers as colnames.
}
\description{
Transforms RNA-seq raw counts to TPM (Transcript Per kilobase
Million).
}
\examples{
tpm <- ic_raw_to_tpm(counts = sample_counts,
                     genes_id = 'ensembl_gene_id',
                     biomart = ensembl_biomart_GRCh38_p13)
}
