% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic_score.R
\name{ic_score}
\alias{ic_score}
\title{ic_score}
\usage{
ic_score(
  tpm,
  metadata,
  response,
  compare = NULL,
  p_label = "p.format",
  colors = c("orange", "black")
)
}
\arguments{
\item{tpm}{Output from the ic_raw_to_tpm function. This is a matrix
containing expression counts as TPM with HGNC gene symbols as rownames
and samples identifiers as colnames.}

\item{metadata}{Data frame that contains supporting variables to the data.}

\item{response}{Unquoted name of the variable indicating the groups to analyse.}

\item{compare}{A character string indicating which method to be used for
comparing means. Options are 't.test' and 'wilcox.test' for two groups or
'anova' and 'kruskal.test' for more groups. Default value is NULL.}

\item{p_label}{Character string specifying label type. Allowed values include
'p.signif' (shows the significance levels), 'p.format' (shows the formatted
p-value).}

\item{colors}{Character vector indicating the colors of the different groups
to compare. Default values are two: black and orange.}
}
\value{
Returns ggplot objects and a data frame.
}
\description{
Calculates and plots immunophenogram (IPG) and immunophenoscores
(IPS) for each sample and each group of study.
}
\examples{
tpm <- ic_raw_to_tpm(counts = sample_counts,
                     genes_id = 'entrezgene_id',
                     biomart = ensembl_biomart_GRCh38_p13)
ips <- ic_score(tpm = tpm,
                metadata = sample_metadata,
                response = MSI_status,
                compare = 'wilcox.test',
                p_label = 'p.format',
                colors = c('orange', 'black'))

}
