% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_ge.R
\name{module_ge}
\alias{module_ge}
\title{module_ge}
\usage{
module_ge(
  counts,
  genes_id,
  metadata,
  colors = c("black", "orange"),
  design,
  ref_level,
  fold_change = 2,
  p.adj = 0.05,
  gmt,
  gsea_pvalue = 0.2
)
}
\arguments{
\item{gsea_pvalue}{}
}
\value{

}
\description{

}
\examples{


}
