% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ge_data_preprocess.R
\name{preprocess_ge_meta}
\alias{preprocess_ge_meta}
\title{preprocess_ge_meta}
\usage{
preprocess_ge_meta(metadata, counts = NULL)
}
\arguments{
\item{metadata}{Data frame that contains supporting variables to the data.}

\item{counts}{Data frame that contains gene expression data as raw counts.}
}
\value{
Generates a data frame.
}
\description{
Prepares metadata file for posterior analysis using DESeq2.
}
\examples{
metadata <- preprocess_ge_meta(metadata = metadata_ge_module,
                               counts = input_ge_module[,-1])
}
